int to size_t

+++ inc/hotkeyPP.h
@@ -200,7 +200,7 @@ namespace HKPP
         template <class T>
         bool VectorEx<T>::Contains(T val)
         {
-            for (int i = 0; i < this->size(); i++)
+            for (size_t i = 0; i < this->size(); i++)
                 if ((*this)[i] == val)
                     return true;
